added reheating and iterations, changed float to int

+++ includes/SudokuEvaluation.h
@@ -4,5 +4,5 @@
 
 int EvaluarColumna(Matrix Sudoku, int NumeroColumna);
 int EvaluarFila(Matrix Sudoku, int NumeroFila);
-float EvaluarGrupos(Matrix Sudoku,numeroGrupoaCeldas GrupoaCeldas,numeroGrupoaSuma GrupoaSuma);
-float EvaluarSudoku(Matrix Sudoku, numeroGrupoaCeldas GrupoaCeldas , numeroGrupoaSuma GrupoaSuma);
+int EvaluarGrupos(Matrix Sudoku,numeroGrupoaCeldas GrupoaCeldas,numeroGrupoaSuma GrupoaSuma);
+int EvaluarSudoku(Matrix Sudoku, numeroGrupoaCeldas GrupoaCeldas , numeroGrupoaSuma GrupoaSuma);
